Update websocketclient.py
""if endpoint not in ["us", "us_quote", "forex", "crypto"]: ""  ,   "us_quote" in this if statement was typed incorrectly, which caused the code  to raise Value Error. The correct is "us-quote" ,   So the final if Statement will be:     ""if endpoint not in ["us", "us-quote", "forex", "crypto"]:"""      , once changed the quote feed will work as expected.  Thanks